- Renames renderDownloadProgress into renderProgress as it can be generally used - Uses renderProgress also for unzipping in ATC, BAG and Kompendium Job
