+++ jobs/atc.js
@@ -31,7 +31,7 @@ var disk = require("../lib/disk");
 var fetchHTML = require("../lib/fetchHTML");
 var parseLink = require("../lib/parseLink");
 var downloadFile = require("../lib/downloadFile");
-var renderDownloadProgress = require("../lib/renderDownloadProgress");
+var renderProgress = require("../lib/renderProgress");
 var readXLSX = require("../lib/atc/readXLSX");
 var addCodes = require("../lib/atc/addCodes");
 var modifyCodes = require("../lib/atc/modifyCodes");
@@ -78,10 +78,10 @@ function atc(done) {
     .then(function (parsedLink) {
       log.timeEnd("ATC", "Parse Link");
       log.time("ATC", "Download");
-      return downloadFile(parsedLink, cfg.download.file, renderDownloadProgress("ATC", "Download"));
+      return downloadFile(parsedLink, cfg.download.file, renderProgress("ATC", "Download"));
     })
     .then(function () {
-      return disk.unzip(cfg.download.file, cfg.download.zipFiles);
+      return disk.unzip(cfg.download.file, cfg.download.zipFiles, renderProgress("ATC", "Unzip"));
     })
     .then(function () {
       log.timeEnd("ATC", "Download");
