+++ lib/disk.js
@@ -51,7 +51,7 @@ var fileWriter = {
               var readStream = streamifier.createReadStream(buffer);
               var inflateStream = zlib.createInflateRaw();
               var writeStream = fs.createWriteStream(dest);
-              var progressStream = progress({length: buffer.length});
+              var progressStream = progress({length: entry.header.size});
 
               readStream.on("error", _reject);
