+++ lib/renderProgress.js
